+++ ext_localconf.php
@@ -21,7 +21,7 @@
 use MindfulMarkup\MindfulA11y\Form\FieldControl\GenerateAltTextControl;
 use MindfulMarkup\MindfulA11y\Form\ValidationErrorDetector;
 use MindfulMarkup\MindfulA11y\Form\FormDataProvider\MetadataAlternativePlaceholderAccessGuard;
-use MindfulMarkup\MindfulA11y\Hooks\DecorativeFileReferenceGuard;
+use MindfulMarkup\MindfulA11y\Hooks\DecorativeFileReferenceDataHandlerGuard;
 use MindfulMarkup\MindfulA11y\Hooks\ScanStateDataHandlerGuard;
 use MindfulMarkup\MindfulA11y\Service\StructureAnalysisTicketService;
 use TYPO3\CMS\Backend\Form\FormDataProvider\TcaColumnsRemoveUnused;
@@ -43,8 +43,8 @@ defined('TYPO3') or die();
 
     $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_tcemain.php']['processDatamapClass'][ScanStateDataHandlerGuard::class]
         = ScanStateDataHandlerGuard::class;
-    $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_tcemain.php']['processDatamapClass'][DecorativeFileReferenceGuard::class]
-        = DecorativeFileReferenceGuard::class;
+    $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['t3lib/class.t3lib_tcemain.php']['processDatamapClass'][DecorativeFileReferenceDataHandlerGuard::class]
+        = DecorativeFileReferenceDataHandlerGuard::class;
 
     $formDataProviders = &$GLOBALS['TYPO3_CONF_VARS']['SYS']['formEngine']['formDataGroup']['tcaDatabaseRecord'];
     $formDataProviders[MetadataAlternativePlaceholderAccessGuard::class] = [